pyproject.toml
[tool.poetry]
name = "news_crawler_python"
version = "0.1.0"
description = ""
authors = ["takeshi0406"]

[tool.poetry.dependencies]
python = "^3.9"
functions-framework = "^3.0.0"
trio = "^0.19.0"
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
python-twitter = "^3.5"
beautifulsoup4 = "^4.10.0"
lxml = "^4.7.1"
httpx = "^0.21.1"
slack-sdk = "^3.13.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = {version = "^21.12b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
